Exception handling in PropertiesAccess needs to be implemented
Implemented ConfigurationException  which is thrown if an IOException occurs when reading from file
Implemented this custom exception which is handled in Application in order to hide the fact that
configuration data is stored on file.  If IOException was handled in Application, it directly ties this class
to a file based configuration format.  This way configuration data could just as easily be fetched from
a database for example.
Refactored Application, PropertiesAccessor and Configuration to enable reading / writing of configuration
for ApplicationMode.STANDALONE_CLIENT or ApplicatinMode.NETWORK_CLIENT